Move ValueList to functions/lib/aggregates for Spark reuse
