+++ velox/exec/tests/SimpleArrayAggAggregate.cpp
@@ -18,7 +18,7 @@
 #include "velox/exec/SimpleAggregateAdapter.h"
 #include "velox/expression/FunctionSignature.h"
 #include "velox/expression/VectorWriters.h"
-#include "velox/functions/prestosql/aggregates/ValueList.h"
+#include "velox/functions/lib/aggregates/ValueList.h"
 
 using namespace facebook::velox::exec;
