+++ velox/functions/lib/aggregates/CMakeLists.txt
@@ -15,7 +15,7 @@
 add_library(
   velox_functions_aggregates
   AverageAggregateBase.cpp CentralMomentsAggregatesBase.cpp
-  SingleValueAccumulator.cpp ValueSet.cpp)
+  SingleValueAccumulator.cpp ValueList.cpp ValueSet.cpp)
 
 target_link_libraries(velox_functions_aggregates velox_exec
                       velox_presto_serializer Folly::folly)
